finish components
